fix: add missing await in rejection tests and deduplicate cli utils
