+++ .thunderbot/cli/repo.ts
@@ -72,3 +72,53 @@ export const getCurrentBranch = async (): Promise<string> => {
   }
   return result.stdout
 }
+
+/** Parse a required argument by flag name from a CLI args array */
+export const parseRequiredArg = (args: string[], flag: string, type: 'number' | 'string'): number | string => {
+  const idx = args.indexOf(flag)
+  if (idx === -1 || idx + 1 >= args.length) {
+    throw new Error(`Missing required argument: ${flag}`)
+  }
+  const raw = args[idx + 1]
+  if (type === 'number') {
+    const num = parseInt(raw, 10)
+    if (isNaN(num)) {
+      throw new Error(`${flag} must be a number, got: ${raw}`)
+    }
+    return num
+  }
+  return raw
+}
+
+/** Parse the --pr flag from args, throwing if missing or non-numeric */
+export const parseRequiredPR = (args: string[]): number =>
+  parseRequiredArg(args, '--pr', 'number') as number
+
+/**
+ * Write a GraphQL query to a temp file and execute it via gh api graphql.
+ * This avoids the $id shell expansion issue that occurs with inline queries.
+ */
+export const runGraphQL = async (query: string, variables: Record<string, string>): Promise<string> => {
+  const { mkdtempSync, writeFileSync, rmSync } = await import('fs')
+  const { join } = await import('path')
+  const { tmpdir } = await import('os')
+
+  const dir = mkdtempSync(join(tmpdir(), 'thunderbot-gql-'))
+  const queryFile = join(dir, 'query.graphql')
+  try {
+    writeFileSync(queryFile, query)
+    const args = ['api', 'graphql', '-F', `query=@${queryFile}`]
+    for (const [key, value] of Object.entries(variables)) {
+      args.push('-f', `${key}=${value}`)
+    }
+    return await gh(args)
+  } finally {
+    rmSync(dir, { recursive: true, force: true })
+  }
+}
+
+/** Run a GraphQL query and parse the JSON result */
+export const runGraphQLJSON = async <T>(query: string, variables: Record<string, string>): Promise<T> => {
+  const raw = await runGraphQL(query, variables)
+  return JSON.parse(raw) as T
+}
